+++ buildprocess/configureWebpack.js
@@ -276,6 +276,7 @@ function configureWebpack(terriaJSBasePath, config, devMode, hot, MiniCssExtract
     config.plugins.push(
         new ForkTsCheckerWebpackPlugin({
             typescript: {
+                memoryLimit:4096,
                 configFile: path.resolve(__dirname, '..', 'tsconfig.json'),
                 diagnosticOptions: {
                     semantic: true,