+++ lib/ModelMixins/GeojsonMixin.ts
@@ -7,7 +7,9 @@ import {
   Geometries,
   Geometry,
   GeometryCollection,
+  MultiPolygon,
   Point,
+  Polygon,
   Properties
 } from "@turf/helpers";
 import i18next from "i18next";
@@ -55,20 +57,17 @@ import filterOutUndefined from "../Core/filterOutUndefined";
 import formatPropertyValue from "../Core/formatPropertyValue";
 import hashFromString from "../Core/hashFromString";
 import isDefined from "../Core/isDefined";
-import { isJsonObject } from "../Core/Json";
+import { isJsonObject, JsonObject, isJsonNumber } from "../Core/Json";
 import { isJson } from "../Core/loadBlob";
 import makeRealPromise from "../Core/makeRealPromise";
 import StandardCssColors from "../Core/StandardCssColors";
-import TerriaError, {
-  networkRequestError,
-  TerriaErrorSeverity
-} from "../Core/TerriaError";
+import TerriaError, { networkRequestError } from "../Core/TerriaError";
 import ProtomapsImageryProvider, {
   GeojsonSource,
   GEOJSON_SOURCE_LAYER_NAME,
   ProtomapsData
-} from "../Map/ProtomapsImageryProvider";
-import Reproject from "../Map/Reproject";
+} from "../Map/ImageryProvider/ProtomapsImageryProvider";
+import Reproject from "../Map/Vector/Reproject";
 import CatalogMemberMixin from "../ModelMixins/CatalogMemberMixin";
 import UrlMixin from "../ModelMixins/UrlMixin";
 import proxyCatalogItemUrl from "../Models/Catalog/proxyCatalogItemUrl";
@@ -77,6 +76,10 @@ import createStratumInstance from "../Models/Definition/createStratumInstance";
 import LoadableStratum from "../Models/Definition/LoadableStratum";
 import Model, { BaseModel } from "../Models/Definition/Model";
 import StratumOrder from "../Models/Definition/StratumOrder";
+import TableStylingWorkflow from "../Models/SelectableDimensions/TableStylingWorkflow";
+import VectorStylingWorkflow from "../Models/SelectableDimensions/VectorStylingWorkflow";
+import { ViewingControl } from "../Models/ViewingControls";
+import Icon from "../Styled/Icon";
 import TableAutomaticStylesStratum from "../Table/TableAutomaticStylesStratum";
 import { GeoJsonTraits } from "../Traits/TraitsClasses/GeoJsonTraits";
 import LegendTraits from "../Traits/TraitsClasses/LegendTraits";
@@ -186,6 +189,13 @@ function GeoJsonMixin<T extends Constructor<Model<GeoJsonTraits>>>(Base: T) {
 
     /** Geojson FeatureCollection in WGS84 */
     @observable.ref _readyData?: FeatureCollectionWithCrs;
+
+    /** Number of features in _readyData FeatureCollection */
+    @observable featureCounts: {
+      point: number;
+      line: number;
+      polygon: number;
+    } = { point: 0, line: 0, polygon: 0 };
 
     constructor(...args: any[]) {
       super(...args);
@@ -398,6 +408,7 @@ function GeoJsonMixin<T extends Constructor<Model<GeoJsonTraits>>>(Base: T) {
     protected async forceLoadMapItems(): Promise<void> {
       let useMvt = this.useMvt;
       const czmlTemplate = this.czmlTemplate;
+      const filterByProperties = this.filterByProperties;
 
       let geoJson: FeatureCollectionWithCrs | undefined;
 
@@ -417,18 +428,55 @@ function GeoJsonMixin<T extends Constructor<Model<GeoJsonTraits>>>(Base: T) {
 
         // Also check for how many features have simply-style properties
         let numFeaturesWithSimpleStyle = 0;
+        const featureCounts = { point: 0, line: 0, polygon: 0 };
 
-        for (let i = 0; i < geoJsonWgs84.features.length; i++) {
-          if (!geoJsonWgs84.features[i].properties) {
-            geoJsonWgs84.features[i].properties = {};
+        const features = geoJsonWgs84.features;
+        // If filtering features - Clear all features and re add them if props match filterByProperties
+        if (filterByProperties) geoJsonWgs84.features = [];
+
+        for (let i = 0; i < features.length; i++) {
+          const feature = features[i];
+          if (!feature.properties) {
+            feature.properties = {};
           }
-          const properties = geoJsonWgs84.features[i].properties!;
+
+          // Filter features by `featureFilterByProps` trait if defined
+          if (filterByProperties) {
+            if (
+              Object.entries(filterByProperties).every(
+                ([key, value]) => feature.properties![key] === value
+              )
+            )
+              geoJsonWgs84.features.push(feature);
+            else continue;
+          }
+
+          const properties = feature.properties!;
           properties[FEATURE_ID_PROP] = i;
+
+          if (
+            feature.geometry.type === "Point" ||
+            feature.geometry.type === "MultiPoint"
+          ) {
+            featureCounts.point++;
+          } else if (
+            feature.geometry.type === "LineString" ||
+            feature.geometry.type === "MultiLineString"
+          ) {
+            featureCounts.line++;
+          } else if (
+            feature.geometry.type === "Polygon" ||
+            feature.geometry.type === "MultiPolygon"
+          ) {
+            featureCounts.polygon++;
+          }
 
           if (useMvt && SIMPLE_STYLE_KEYS.find(key => properties[key])) {
             numFeaturesWithSimpleStyle++;
           }
         }
+
+        runInAction(() => (this.featureCounts = featureCounts));
 
         // If more than 50% of features have simple style properties - disable table styling
         if (numFeaturesWithSimpleStyle / geoJsonWgs84.features.length >= 0.5) {
@@ -646,7 +694,7 @@ function GeoJsonMixin<T extends Constructor<Model<GeoJsonTraits>>>(Base: T) {
                 this.stylesWithDefaults.markerColor.toCssColorString()
               ),
               width: this.stylesWithDefaults.markerStrokeWidth,
-              stroke: this.stylesWithDefaults.stroke.toCssColorString(),
+              stroke: this.stylesWithDefaults.markerStroke.toCssColorString(),
               opacity: this.stylesWithDefaults.markerOpacity
             }),
             minzoom: 0,
@@ -677,30 +725,92 @@ function GeoJsonMixin<T extends Constructor<Model<GeoJsonTraits>>>(Base: T) {
         }
       ];
 
-      // Create a czml packet for each geoJson Point feature
-      // Set czml position (cartographicDegrees) to point coordinates
+      // Create a czml packet for each geoJson Point/Polygon feature
+      // For point: set czml position (CartographicDegrees) to point coordinates
+      // For polygon: set czml positions array (CartographicDegreesListValue) for the `polygon` property
+
       // Set czml properties to feature properties
       for (let i = 0; i < geoJson.features.length; i++) {
         const feature = geoJson.features[i];
-        if (feature !== null && feature.geometry?.type === "Point") {
-          const point = feature.geometry as Point;
+        if (feature === null || feature.geometry.type === "Line") {
+          continue;
+        }
+
+        if (feature.geometry?.type === "Point") {
           const czml = clone(czmlTemplate ?? {}, true);
+
+          const point = feature.geometry as Point;
           const coords = point.coordinates;
+
+          // Add height = 0 if no height provided
           if (coords.length === 2) {
             coords[2] = 0;
           }
+
+          if (isJsonNumber(this.czmlTemplate?.heightOffset)) {
+            coords[2] += this.czmlTemplate!.heightOffset;
+          }
+
           czml.position = {
             cartographicDegrees: point.coordinates
           };
 
-          if (feature.properties !== null) {
+          czml.properties = Object.assign(
+            czml.properties ?? {},
+            stringifyFeatureProperties(feature.properties ?? {})
+          );
+          rootCzml.push(czml);
+        } else if (
+          feature.geometry?.type === "Polygon" ||
+          (feature.geometry?.type === "MultiPolygon" && czmlTemplate?.polygon)
+        ) {
+          const czml = clone(czmlTemplate ?? {}, true);
+
+          // To handle both Polygon and MultiPolygon - transform Polygon coords into MultiPolygon coords
+          const multiPolygonGeom =
+            feature.geometry?.type === "Polygon"
+              ? [(feature.geometry as Polygon).coordinates]
+              : (feature.geometry as MultiPolygon).coordinates;
+
+          // Loop through Polygons in MultiPolygon
+          for (let j = 0; j < multiPolygonGeom.length; j++) {
+            const geom = multiPolygonGeom[j];
+            const positions: number[] = [];
+            const holes: number[][] = [];
+
+            geom[0].forEach(coords => {
+              if (isJsonNumber(this.czmlTemplate?.heightOffset)) {
+                coords[2] = (coords[2] ?? 0) + this.czmlTemplate!.heightOffset;
+              }
+              positions.push(coords[0], coords[1], coords[2]);
+            });
+
+            geom.forEach((ring, idx) => {
+              if (idx === 0) return;
+
+              holes.push(
+                ring.reduce<number[]>((acc, current) => {
+                  if (isJsonNumber(this.czmlTemplate?.heightOffset)) {
+                    current[2] =
+                      (current[2] ?? 0) + this.czmlTemplate!.heightOffset;
+                  }
+
+                  acc.push(current[0], current[1], current[2]);
+
+                  return acc;
+                }, [])
+              );
+            });
+
+            czml.polygon.positions = { cartographicDegrees: positions };
+            czml.polygon.holes = { cartographicDegrees: holes };
+
             czml.properties = Object.assign(
               czml.properties ?? {},
-              feature.properties
+              stringifyFeatureProperties(feature.properties ?? {})
             );
+            rootCzml.push(czml);
           }
-
-          rootCzml.push(czml);
         }
       }
 
@@ -710,9 +820,10 @@ function GeoJsonMixin<T extends Constructor<Model<GeoJsonTraits>>>(Base: T) {
     @computed
     get stylesWithDefaults() {
       const defaults = {
-        markerSize: 20,
+        markerSize: 24,
         markerColor: getRandomCssColor(this.name ?? ""),
         stroke: getColor(this.terria.baseMapContrastColor),
+        markerStroke: getColor(this.terria.baseMapContrastColor),
         polygonStroke: getColor(this.terria.baseMapContrastColor),
         polylineStroke: getRandomCssColor(this.name ?? ""),
         markerStrokeWidth: 1,
@@ -737,12 +848,20 @@ function GeoJsonMixin<T extends Constructor<Model<GeoJsonTraits>>>(Base: T) {
           defaults.markerColor
         ),
         stroke: defaultColor(this.style.stroke, defaults.stroke),
-        polygonStroke: defaultColor(this.style.stroke, defaults.polygonStroke),
-        polylineStroke: defaultColor(
-          this.style.stroke,
-          defaults.polylineStroke
+        polygonStroke: defaultColor(
+          this.style["polygon-stroke"] ?? this.style.stroke,
+          defaults.polygonStroke
         ),
         // Note these specific stroke widths are only used for geojson-vt
+
+        polylineStroke: defaultColor(
+          this.style["polyline-stroke"] ?? this.style.stroke,
+          defaults.polylineStroke
+        ),
+        markerStroke: defaultColor(
+          this.style["marker-stroke"] ?? this.style.stroke,
+          defaults.markerStroke
+        ),
         markerStrokeWidth:
           this.style["marker-stroke-width"] ??
           this.style["stroke-width"] ??
@@ -768,6 +887,7 @@ function GeoJsonMixin<T extends Constructor<Model<GeoJsonTraits>>>(Base: T) {
         options.stroke.alpha = this.style["stroke-opacity"];
         options.polygonStroke.alpha = this.style["stroke-opacity"];
         options.polylineStroke.alpha = this.style["stroke-opacity"];
+        options.markerStroke.alpha = this.style["stroke-opacity"];
       }
 
       if (isDefined(this.style["fill-opacity"])) {
@@ -1041,6 +1161,46 @@ function GeoJsonMixin<T extends Constructor<Model<GeoJsonTraits>>>(Base: T) {
      */
     async forceLoadTableData() {
       return undefined;
+    }
+
+    @computed get viewingControls(): ViewingControl[] {
+      return this.useMvt && this.activeStyle // Note we want falsy here for activeStyle ("" is equivalent to undefined)
+        ? [
+            // We replace the TableStylingWorkflow with VectorStylingWorkflow
+            ...super.viewingControls.filter(
+              v => v.id !== TableStylingWorkflow.type
+            ),
+            {
+              id: VectorStylingWorkflow.type,
+              name: "Edit Style",
+              onClick: viewState => {
+                runInAction(() => {
+                  viewState.terria.selectableDimensionWorkflow = new VectorStylingWorkflow(
+                    this
+                  );
+                });
+              },
+              icon: { glyph: Icon.GLYPHS.layers }
+            }
+          ]
+        : [];
+    }
+
+    /** This is a temporary button which shows in the Legend in the Workbench, if custom styling has been applied. */
+    @computed get legendButton() {
+      // Show the button if TableMixin.legendButton is showing - or we have values in style User stratum
+      return super.legendButton ||
+        filterOutUndefined(Object.values(this.style.strata.get("user") ?? {}))
+          .length > 0
+        ? {
+            title: "Custom",
+            onClick: action(() => {
+              this.terria.selectableDimensionWorkflow = new VectorStylingWorkflow(
+                this
+              );
+            })
+          }
+        : undefined;
     }
   }
   return GeoJsonMixin;
@@ -1457,7 +1617,7 @@ export function getColor(color: String | string | Color): Color {
   }
 }
 
-function parseMarkerSize(sizeString?: string): number | undefined {
+export function parseMarkerSize(sizeString?: string): number | undefined {
   const sizes: { [name: string]: number } = {
     small: 24,
     medium: 48,
@@ -1472,4 +1632,22 @@ function parseMarkerSize(sizeString?: string): number | undefined {
     return sizes[sizeString];
   }
   return parseInt(sizeString, 10); // SimpleStyle doesn't allow 'marker-size: 20', but people will do it.
+}
+
+function stringifyFeatureProperties(featureProps: JsonObject | undefined) {
+  return Object.keys(featureProps ?? {}).reduce<{
+    [key: string]: string;
+  }>((properties, key) => {
+    const featureProp = featureProps![key];
+    if (typeof featureProp === "string") {
+      properties[key] = featureProp;
+    } else if (
+      isDefined(featureProp) &&
+      featureProp !== null &&
+      typeof featureProp.toString === "function"
+    )
+      properties[key] = featureProp.toString();
+
+    return properties;
+  }, {});
 }