+++ lib/ReactViews/BottomDock/BottomDock.jsx
@@ -1,5 +1,3 @@
-"use strict";
-
 import createReactClass from "create-react-class";
 import { runInAction } from "mobx";
 import { observer } from "mobx-react";
@@ -10,6 +8,7 @@ import ChartPanel from "../Custom/Chart/ChartPanel";
 import Styles from "./bottom-dock.scss";
 import ChartDisclaimer from "./ChartDisclaimer";
 import Timeline from "./Timeline/Timeline";
+import withControlledVisibility from "../HOCs/withControlledVisibility";
 
 const BottomDock = observer(
   createReactClass({
@@ -60,21 +59,25 @@ const BottomDock = observer(
             background: ${p => p.theme.dark};
           `}
         >
+          <div id="TJS-BottomDockFirstPortal" />
           <ChartDisclaimer terria={terria} viewState={this.props.viewState} />
           <ChartPanel
             terria={terria}
             onHeightChange={this.onHeightChange}
             viewState={this.props.viewState}
           />
-          <If condition={top}>
-            <Timeline terria={terria} />
-          </If>
+          {top && (
+            <Timeline
+              terria={terria}
+              elementConfig={this.props.terria.elements.get("timeline")}
+            />
+          )}
           {/* Used for react portals - do not remove without updating portals using this */}
-          <div id="TJS-BottomDockPortalForTool" />
+          <div id="TJS-BottomDockLastPortal" />
         </div>
       );
     }
   })
 );
 
-module.exports = measureElement(BottomDock, false);
+export default withControlledVisibility(measureElement(BottomDock, false));