+++ lib/Styled/Icon.tsx
@@ -1,6 +1,6 @@
+import classNames from "classnames";
 import React from "react";
 import styled from "styled-components";
-import classNames from "classnames";
 
 // Icon
 export const GLYPHS = {
@@ -130,10 +130,16 @@ export const GLYPHS = {
   maximize: require("../../wwwroot/images/icons/maximize.svg"),
   closeTool: require("../../wwwroot/images/icons/close-tool.svg"),
   moreItems: require("../../wwwroot/images/icons/more-items.svg"),
-  info: require("../../wwwroot/images/icons/info.svg")
+  info: require("../../wwwroot/images/icons/info.svg"),
+  leftSmall: require("../../wwwroot/images/icons/left-small.svg"),
+  rightSmall: require("../../wwwroot/images/icons/right-small.svg"),
+  compareLeftPanel: require("../../wwwroot/images/icons/compare-left-panel.svg"),
+  compareRightPanel: require("../../wwwroot/images/icons/compare-right-panel.svg"),
+  compareBothPanels: require("../../wwwroot/images/icons/compare-both-panels.svg"),
+  closeCircle: require("../../wwwroot/images/icons/close-circle.svg")
 };
 
-interface IconProps {
+export interface IconProps {
   glyph: { id: string };
   style?: any;
   className?: string;
@@ -152,7 +158,7 @@ export const Icon: React.FC<IconProps> = (props: IconProps) => {
   );
 };
 
-interface IStyledIconProps {
+export interface IStyledIconProps {
   displayInline?: boolean;
   styledWidth?: string;
   styledHeight?: string;