+++ lib/Traits/TraitsClasses/MapboxStyleCatalogItemTraits.ts
@@ -3,10 +3,10 @@ import mixTraits from "../mixTraits";
 import CatalogMemberTraits from "./CatalogMemberTraits";
 import LegendOwnerTraits from "./LegendOwnerTraits";
 import MappableTraits from "./MappableTraits";
-import RasterLayerTraits from "./RasterLayerTraits";
+import ImageryProviderTraits from "./ImageryProviderTraits";
 
 export default class MapboxStyleCatalogItemTraits extends mixTraits(
-  RasterLayerTraits,
+  ImageryProviderTraits,
   CatalogMemberTraits,
   MappableTraits,
   LegendOwnerTraits
@@ -53,20 +53,4 @@ export default class MapboxStyleCatalogItemTraits extends mixTraits(
     type: "boolean"
   })
   scaleFactor = false;
-
-  @primitiveTrait({
-    name: "scaleFactor",
-    description:
-      "The minimum level-of-detail supported by the imagery provider. Take care when specifying this that the number of tiles at the minimum level is small, such as four or less. A larger number is likely to result in rendering problems.",
-    type: "number"
-  })
-  minimumLevel = 0;
-
-  @primitiveTrait({
-    name: "scaleFactor",
-    description:
-      "The maximum level-of-detail supported by the imagery provider, or undefined if there is no limit.",
-    type: "number"
-  })
-  maximumLevel?: number;
 }