+++ lib/Traits/TraitsClasses/WebMapTileServiceCatalogItemTraits.ts
@@ -12,7 +12,7 @@ import LayerOrderingTraits from "./LayerOrderingTraits";
 import LegendOwnerTraits from "./LegendOwnerTraits";
 import LegendTraits from "./LegendTraits";
 import MappableTraits from "./MappableTraits";
-import RasterLayerTraits from "./RasterLayerTraits";
+import ImageryProviderTraits from "./ImageryProviderTraits";
 import UrlTraits from "./UrlTraits";
 
 export class WebMapTileServiceAvailableStyleTraits extends ModelTraits {
@@ -73,7 +73,7 @@ export default class WebMapServiceCatalogItemTraits extends mixTraits(
   FeatureInfoTraits,
   LayerOrderingTraits,
   GetCapabilitiesTraits,
-  RasterLayerTraits,
+  ImageryProviderTraits,
   UrlTraits,
   MappableTraits,
   CatalogMemberTraits,