+++ CHANGES.md
@@ -11,6 +11,61 @@ Change Log
 * Added experimental routing system - there may be breaking changes to this system in subsequent patch releases for a short time. The routes currently include:
   * `/story/:share-id` ➡ loads share JSON from a URL `${configParameters.storyRouteUrlPrefix}:share-id` (`configParameters.storyRouteUrlPrefix` must have a trailing slash)
   * `/catalog/:id` ➡ opens the data catalogue to the specified member
+* Fixed a polyline position update bug in `LeafletVisualizer`. Polylines with time varying position will now correctly animate in leaflet mode.
+* Change button cursor to pointer
+* Add `GeoJsonTraits.filterByProperties` - this can be used to filter GeoJSON features by properties
+* Add GeoJSON `czmlTemplate` support for `Polygon/MultiPolygon`
+* Add custom `heightOffset` property to `czmlTemplate`
+* Fixed a bug where Cesium3DTilePointFeature info is not shown when being clicked.
+* Added optional `onDrawingComplete` callback to `UserDrawing` to receive drawn points or rectangle when the drawing is complete.
+* Fixed a bug in `BoxDrawing` where the box can be below ground after initialization even when setting `keepBoxAboveGround` to true.
+* Fix `viewCatalogMember` bug - where `_previewItem` was being set too late.
+* Improve error message in `DataPreview` for references.
+* Fix alignment of elements in story panel and move some styling from scss to styled components
+* Click on the stories button opens a story builder (button on the left from story number)
+* Added ASGS 2021 regions to region mapping:
+  * SA1-4 (e.g. sa3_code_2021)
+  * GCCSA
+  * STE & AUS (aliased to existing 2011/2016 data due to no change in geometry, names & codes)
+* Added LGA regions from 2019 & 2021 to region mapping - only usable by lga code
+* Increase `ForkTsCheckerWebpackPlugin` memoryLimit to 4GB
+* Add `renderInline` option to markdownToHtml/React + TSify files
+* Organise `lib/Map` into folder structure
+* When `modelDimensions` are changed, `loadMapItems()` is automatically called
+* Add `featureCounts` to `GeoJsonMixin` - this tracks number of GeoJSON Features by type
+* Add `polygon-stroke`, `polyline-stroke` and `marker-stroke` to GeoJSON `StyleTraits` - these are only applied to geojson-vt features (not Cesium Primitives)
+* TableMixin manual region mapping dimensions are now in a `SelectableDimensionGroup`
+* Fix misc font/color styles
+* Create reusable `StyledTextArea` component
+* `Collapsible` improvements:
+  * Add `"checkbox"` `btnStyle`
+  * `onToggle` can now stop event propagation
+  * `title` now supports custom markdown
+* Add `rightIcon` and `textLight` props to `Button`
+* New `addTerriaScrollbarStyles` scss mixin
+* `TableAutomaticStylesStratum` now creates `styles` for every column - but will hide columns depending on `TableColumnType`
+* `TableAutomaticStylesStratum.numberFormatOptions` is now `TableStyle.numberFormatOptions`
+* Implement `TableColorStyleTraits.legendTicks` - this will determine number of ticks for `ContinuousColorMap` legends
+* `DiscreteColorMap` will now use `minimumValue`/`maximumValue` to calculate bins
+* `SelectableDimensions` improvements
+  * Add `color`, `text`, `numeric` and `button` types
+  * Add `onToggle` function to `SelectableDimensionGroup`
+  * `Group` and `CheckboxGroup` now share the same UI and use `Collapsible`
+  * `enum` (previously `select`) now uses `react-select` component
+  * `color` uses `react-color` component
+  * `DimensionSelectorSection` / `DimensionSelector*` are now named the same as the model - eg `SelectableDimension`
+* Create `Portal`, `PortalContainer`,`SidePanelContainer` and `WorkflowPanelContainer`. There are used by `WorkflowPanel`.
+* Create `WorkflowPanel` - a basic building block for creating Workflows that sit on top of the workbench
+  * It has three reusable components, `Panel`, `PanelButton`, `PanelMenu`
+* Create `selectableDimensionWorkflow` - This uses `WorkflowPanel` to show `SelectableDimensions` in a separate side panel.
+  * `TableStylingWorkflow` - set styling options for TableMixin models
+  * `VectorStylingWorkflow` - this extends `TableStylingWorkflow` - used to set styling options for GeoJsonMixin models (for Protomaps/geojson-vt only)
+* Create `viewingControls` concept. This can be used to add menu items to workbench items menu (eg "Remove", "Export", ...)
+  * TSXify `ViewingControls`
+* Add temporary `legendButton` property - this is used to show a "Custom" button above the Legend if custom styling has been applied
+  * This uses new `TableStyle.isCustom` property
+* Move workbench item controls from `WorkbenchItem.jsx` `WorkbenchItemControls.tsx`
+* Add `UrlTempalteImageryCatalogItem`, rename `RasterLayerTraits` to `ImageryProviderTraits` and add some properties.
 * [The next improvement]
 
 #### 8.1.25 - 2022-03-16